# python-flask/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "NGSI V2 API"
  version: "v2"
  title: "ngsi-v2"
  license:
    name: "MIT"
    url: "http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT"
host: "https://api.s.orchestracities.com/context/v2/"
basePath: "/v2"
schemes:
- "https"
produces:
- "application/json"
security:
- Bearer: []
paths:
...
  /entities:
    get:
      tags:
      - "Entities"
      description: "Retrieves a list of entities that match different criteria by\
        \ id, type, pattern matching (either id or type)\nand/or those which match\
        \ a query or geographical query (see [Simple Query Language](#simple_query_language)\
        \ and \n[Geographical Queries](#geographical_queries)). A given entity has\
        \ to match all the criteria to be retrieved\n(i.e., the criteria is combined\
        \ in a logical AND way). Note that pattern matching query parameters are incompatible\n\
        (i.e. mutually exclusive) with their corresponding exact matching parameters,\
        \ i.e. `idPattern` with `id` and\n`typePattern` with `type`.\nThe response\
        \ payload is an array containing one object per matching entity. Each entity\
        \ follows\nthe JSON entity Representation format (described in \"JSON Entity\
        \ Representation\" section).\n\nResponse code:\n* Successful operation uses\
        \ 200 OK\n* Errors use a non-2xx and (optionally) an error payload. See subsection\
        \ on \"Error Responses\" for\n  more details."
      operationId: "list_entities"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "query"
        description: "A comma-separated list of elements.\nRetrieve entities whose\
          \ ID matches one of the elements in the list.\nIncompatible with idPattern."
        required: false
        type: "string"
      - name: "type"
        in: "query"
        description: "comma-separated list of elements.\nRetrieve entities whose type\
          \ matches one of the elements in the list.\nIncompatible with `typePattern`."
        required: false
        type: "string"
      - name: "idPattern"
        in: "query"
        description: "A correctly formated regular expression.\nRetrieve entities\
          \ whose ID matches the regular expression. Incompatible with id."
        required: false
        type: "string"
      - name: "typePattern"
        in: "query"
        description: "A correctly formated regular expression.\nRetrieve entities\
          \ whose type matches the regular expression. Incompatible with `type`."
        required: false
        type: "string"
      - name: "q"
        in: "query"
        description: "A query expression, composed of a list of statements\nseparated\
          \ by `;`, i.e., q=statement;statements;statement.\nSee [Simple Query Language\
          \ specification](#simple_query_language)."
        required: false
        type: "string"
      - name: "mq"
        in: "query"
        description: "A query expression for attribute metadata,\ncomposed of a list\
          \ of statements separated by `;`, i.e., mq=statement1;statement2;statement3.\n\
          See [Simple Query Language specification](#simple_query_language)."
        required: false
        type: "string"
      - name: "georel"
        in: "query"
        description: "Spatial relationship between matching entities and a\nreference\
          \ shape. See [Geographical Queries](#geographical_queries)."
        required: false
        type: "string"
      - name: "geometry"
        in: "query"
        description: "Geografical area to which the query is restricted.\nSee [Geographical\
          \ Queries](#geographical_queries)."
        required: false
        type: "string"
      - name: "coords"
        in: "query"
        description: "List of latitude-longitude\npairs of coordinates separated by\
          \ ';'. See [Geographical Queries](#geographical_queries)."
        required: false
        type: "string"
      - name: "limit"
        in: "query"
        description: "Limits the number of entities to be retrieved"
        required: false
        type: "number"
        format: "double"
      - name: "offset"
        in: "query"
        description: "Establishes the offset from where entities are retrieved"
        required: false
        type: "number"
        format: "double"
      - name: "attrs"
        in: "query"
        description: "Comma-separated list of attribute names whose data\nare to be\
          \ included in the response.\nThe attributes are retrieved in the order specified\
          \ by this parameter. If this parameter is\nnot included, the attributes\
          \ are retrieved in arbitrary order."
        required: false
        type: "string"
      - name: "orderBy"
        in: "query"
        description: "Criteria for ordering results.\nSee \"Ordering Results\" section\
          \ for details."
        required: false
        type: "string"
      - name: "options"
        in: "query"
        description: "Options dictionary"
        required: false
        type: "string"
        enum:
        - "count"
        - "keyValues"
        - "values"
        - "unique"
      responses:
        200:
          description: "JSON Entity Representation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Entity"
        default:
          description: "error payload"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "swagger_server.controllers.entities_controller"
    post:
      tags:
      - "Entities"
      description: "The payload is an object representing the entity to be created.\
        \ The object follows\nthe JSON entity Representation format (described in\
        \ a \"JSON Entity Representation\" section).\nResponse:\n* Successful operation\
        \ uses 201 Created. Reponse includes a `Location` header with the URL of the\n\
        \  created entity.\n* Errors use a non-2xx and (optionally) an error payload.\
        \ See subsection on \"Error Responses\" for\n  more details."
      operationId: "create_entity"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "options"
        in: "query"
        description: "Options dictionary"
        required: false
        type: "string"
        enum:
        - "keyValues"
      - in: "body"
        name: "body"
        description: "JSON Entity Representation"
        required: true
        schema:
          $ref: "#/definitions/Entity"
      responses:
        200:
          description: ""
        default:
          description: "error payload"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "swagger_server.controllers.entities_controller"
  /entities/{entityId}:
    get:
      tags:
      - "Entities"
      description: "The response is an object representing the entity identified by\
        \ the ID. The object follows\nthe JSON entity Representation format (described\
        \ in \"JSON Entity Representation\" section).\nThis operation must return\
        \ one entity element only, but there may be more than one entity with the\n\
        same ID (e.g. entities with same ID but different types).\nIn such case, an\
        \ error message is returned, with the HTTP status code set to 409 Conflict.\n\
        Response:\n* Successful operation uses 200 OK\n* Errors use a non-2xx and\
        \ (optionally) an error payload. See subsection on \"Error Responses\" for\
        \ more details."
      operationId: "retrieve_entity"
      produces:
      - "application/json"
      parameters:
      - name: "entityId"
        in: "path"
        description: "Id of the entity to be retrieved"
        required: true
        type: "string"
      - name: "type"
        in: "query"
        description: "Entity type, to avoid ambiguity in case there are several\n\
          entities with the same entity id."
        required: false
        type: "string"
      - name: "attrs"
        in: "query"
        description: "Comma-separated list of attribute names whose\ndata must be\
          \ included in the response. The attributes are retrieved in the order specified\
          \ by\nthis parameter.\nIf this parameter is not included, the attributes\
          \ are retrieved in arbitrary order, and all\nthe attributes of the entity\
          \ are included in the response."
        required: false
        type: "string"
      - name: "options"
        in: "query"
        description: "Options dictionary"
        required: false
        type: "string"
        enum:
        - "keyValues"
        - "values"
        - "unique"
      responses:
        200:
          description: "JSON Entity Representation"
          schema:
            $ref: "#/definitions/Entity"
        default:
          description: "error payload"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-unitTests: []
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.entities_controller"
    delete:
      tags:
      - "Entities"
      description: "Delete the entity.\nResponse:\n* Successful operation uses 204\
        \ No Content\n* Errors use a non-2xx and (optionally) an error payload. See\
        \ subsection on \"Error Responses\" for\n  more details."
      operationId: "remove_entity"
      produces:
      - "application/json"
      parameters:
      - name: "entityId"
        in: "path"
        description: "Id of the entity to be deleted"
        required: true
        type: "string"
      - name: "type"
        in: "query"
        description: "Entity type, to avoid ambiguity in the case there are several\n\
          entities with the same entity id."
        required: false
        type: "string"
      responses:
        200:
          description: ""
        default:
          description: "error payload"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-unitTests: []
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.entities_controller"
  /entities/{entityId}/attrs:
    get:
      tags:
      - "Entities"
      description: "This request is similar to retreiving the whole entity, however\
        \ this one omits the `id` and `type`\nfields.\nJust like the general request\
        \ of getting an entire entity, this operation must return only one\nentity\
        \ element. If more than one entity with the same ID is found (e.g. entities\
        \ with\nsame ID but different type), an error message is returned, with the\
        \ HTTP status code set to\n409 Conflict.\nResponse:\n* Successful operation\
        \ uses 200 OK\n* Errors use a non-2xx and (optionally) an error payload. See\
        \ subsection on \"Error Responses\" for\n  more details."
      operationId: "retrieve_entity_attributes"
      produces:
      - "application/json"
      parameters:
      - name: "entityId"
        in: "path"
        description: "Id of the entity to be retrieved"
        required: true
        type: "string"
      - name: "type"
        in: "query"
        description: "Entity type, to avoid ambiguity in the case there are several\n\
          entities with the same entity id."
        required: false
        type: "string"
      - name: "attrs"
        in: "query"
        description: "Comma-separated list of attribute names whose\ndata are to be\
          \ included in the response. The attributes are retrieved in the order specified\n\
          by this parameter.\nIf this parameter is not included, the attributes are\
          \ retrieved in arbitrary order, and all\nthe attributes of the entity are\
          \ included in the response."
        required: false
        type: "string"
      - name: "options"
        in: "query"
        description: "Options dictionary"
        required: false
        type: "string"
        enum:
        - "keyValues"
        - "values"
        - "unique"
      responses:
        200:
          description: "JSON Attribute Representation"
          schema:
            $ref: "#/definitions/Attribute"
        default:
          description: "error payload"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "swagger_server.controllers.entities_controller"
    post:
      tags:
      - "Entities"
      description: "The request payload is an object representing the attributes to\
        \ append or update. The object follows\nthe JSON entity Representation format\
        \ (described in \"JSON Entity Representation\" section), except\nthat `id`\
        \ and `type` are not allowed.\nThe entity attributes are updated with the\
        \ ones in the payload, depending on\nwhether the `append` operation option\
        \ is used or not.\n* If `append` is not used: the entity attributes are updated\
        \ (if they previously exist) or appended\n  (if they don't previously exist)\
        \ with the ones in the payload.\n* If `append` is used (i.e. strict append\
        \ semantics): all the attributes in the payload not\n  previously existing\
        \ in the entity are appended. In addition to that, in case some of the\n \
        \ attributes in the payload already exist in the entity, an error is returned.\n\
        Response:\n* Successful operation uses 204 No Content\n* Errors use a non-2xx\
        \ and (optionally) an error payload. See subsection on \"Error Responses\"\
        \ for\n  more details."
      operationId: "update_or_append_entity_attributes"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "entityId"
        in: "path"
        description: "Entity id to be updated"
        required: true
        type: "string"
      - name: "type"
        in: "query"
        description: "Entity type, to avoid ambiguity in case there are several\n\
          entities with the same entity id."
        required: false
        type: "string"
      - name: "options"
        in: "query"
        description: "Operations options"
        required: false
        type: "string"
        enum:
        - "append"
        - "keyValues"
      - in: "body"
        name: "body"
        description: "JSON Attribute Representation"
        required: true
        schema:
          $ref: "#/definitions/Attribute"
      responses:
        200:
          description: ""
        default:
          description: "error payload"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "swagger_server.controllers.entities_controller"
    put:
      tags:
      - "Entities"
      description: "The request payload is an object representing the new entity attributes.\
        \ The object follows\nthe JSON entity Representation format (described in\
        \ a \"JSON Entity Representation\" above), except\nthat `id` and `type` are\
        \ not allowed.\nThe attributes previously existing in the entity are removed\
        \ and replaced by the ones in the\nrequest.\nResponse:\n* Successful operation\
        \ uses 204 No Content\n* Errors use a non-2xx and (optionally) an error payload.\
        \ See subsection on \"Error Responses\" for\n  more details."
      operationId: "replace_all_entity_attributes"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "entityId"
        in: "path"
        description: "Id of the entity in question."
        required: true
        type: "string"
      - name: "type"
        in: "query"
        description: "Entity type, to avoid ambiguity in the case there are several\n\
          entities with the same entity id."
        required: false
        type: "string"
      - name: "options"
        in: "query"
        description: "Operations options"
        required: false
        type: "string"
        enum:
        - "keyValues"
      - in: "body"
        name: "body"
        description: "JSON Attribute Representation"
        required: true
        schema:
          $ref: "#/definitions/Attribute"
      responses:
        200:
          description: ""
        default:
          description: "error payload"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "swagger_server.controllers.entities_controller"
    patch:
      tags:
      - "Entities"
      description: "The request payload is an object representing the attributes to\
        \ update. The object follows\nthe JSON entity Representation format (described\
        \ in \"JSON Entity Representation\" section), except\nthat `id` and `type`\
        \ are not allowed.\nThe entity attributes are updated with the ones in the\
        \ payload. In addition to that, if one or more\nattributes in the payload\
        \ doesn't exist in the entity, an error is returned.\nResponse:\n* Successful\
        \ operation uses 204 No Content\n* Errors use a non-2xx and (optionally) an\
        \ error payload. See subsection on \"Error Responses\" for\n  more details."
      operationId: "update_existing_entity_attributes"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "entityId"
        in: "path"
        description: "Id of the entity to be updated"
        required: true
        type: "string"
      - name: "type"
        in: "query"
        description: "Entity type, to avoid ambiguity in case there are several\n\
          entities with the same entity id."
        required: false
        type: "string"
      - name: "options"
        in: "query"
        description: "Operations options"
        required: false
        type: "string"
        enum:
        - "keyValues"
      - in: "body"
        name: "body"
        description: "JSON Attribute Representation"
        required: true
        schema:
          $ref: "#/definitions/Attribute"
      responses:
        200:
          description: ""
        default:
          description: "error payload"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "swagger_server.controllers.entities_controller"
...
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  ErrorResponse:
    type: "object"
    required:
    - "error"
    properties:
      error:
        type: "string"
      description:
        type: "string"
    description: "All NGSIv2 server implementations must use the following HTTP response\
      \ codes and error texts. However, the particular text used for description field\
      \ is an implementation specific aspect. Error list (HTTP response code in parenthesis):\
      \ * `ParseError` (400). The incoming JSON payload cannot be parsed. * `BadRequest`\
      \ (400). The incoming request is invalid in this context. * `NotFound` (404).\
      \ The resource (entity, subscription, etc.) referred in the request has not\
      \ been found. * `TooManyResults` (409). There are several results that match\
      \ with the resource identification used in the request. * `ContentLengthRequired`\
      \ (411). Zero/No Content-Length in PUT/POST/PATCH request. * `RequestEntityTooLarge`\
      \ (413). Payload is too large. * `UnsupportedMediaType` (415). Request content\
      \ type is not supported. * `InvalidModification` (422). Some piece of information\
      \ is missing in payload. * `NotSupportedQuery` (422). The implementation does\
      \ not support the query issued. * `NoResourcesAvailable` (413). There are no\
      \ server resources to fulfill the client request. New error codes may be defined\
      \ in new iterations of this specificatio."
    example:
      error: "operation not supported"
      description: "this implementation of NGSI v2 does not support the called operation"
...
  Attribute:
    type: "object"
    required:
    - "metadata"
    - "type"
    - "value"
    properties:
      value:
        type: "object"
        properties: {}
      type:
        type: "string"
      metadata:
        type: "object"
        properties: {}
    example:
      value: 35.6
      type: "Number"
      metadata:
        accuracy:
          value: 2
          type: "Number"
...
  Entity:
    type: "object"
    required:
    - "id"
    - "type"
    properties:
      type:
        type: "string"
      id:
        type: "string"
    example:
      type: "Room"
      id: "DC_S1-D41"
      temperature:
        value: 35.6
        type: "Number"
        metadata:
          accuracy:
            value: 2
            type: "Number"
...